audio syncronization is now perfect

+++ src/Game_Consts.h
@@ -2,12 +2,12 @@
 #ifndef GAMECONSTS_H_
 #define GAMECONSTS_H_
 
-const unsigned int ScreenWidth = 1024;
-const unsigned int ScreenHeight = 768;
-const unsigned int PlayfieldWidth = 800;
-const unsigned int PlayfieldHeight = 600;
-const unsigned int CircleSize = 80;
-const unsigned int ScreenOffset = 80;
+const uint16 ScreenWidth = 1024;
+const uint16 ScreenHeight = 768;
+const uint16 PlayfieldWidth = 800;
+const uint16 PlayfieldHeight = 600;
+const uint16 CircleSize = 80;
+const int16 ScreenOffset = 80;
 const float		   LeniencyHitTime = 0.35f;
 const float		   HoldLeniencyHitTime = 0.1f;
 
@@ -25,14 +25,14 @@ enum Judgement
 
 struct EvaluationData
 {
-	unsigned int MaxCombo;
-	unsigned int NumNG;
-	unsigned int NumOK;
-	unsigned int NumMisses;
-	unsigned int NumBads;
-	unsigned int NumGreats;
-	unsigned int NumPerfects;
-	unsigned int NumExcellents;
+	uint32 MaxCombo;
+	uint32 NumNG;
+	uint32 NumOK;
+	uint32 NumMisses;
+	uint32 NumBads;
+	uint32 NumGreats;
+	uint32 NumPerfects;
+	uint32 NumExcellents;
 };
 
 float _ScreenDifference();
